prep for tutoring

+++ src/handler/AutoException.java
@@ -38,7 +38,7 @@ public class AutoException extends Exception {
 		ExceptionRegistry.finalizeAllPendingRegistrars();
 	}
 
-	public long getExceptionID(){
+	public short getExceptionID(){
 		return ExceptionID;
 	}
